pyproject.toml
[tool.poetry]
name = "vtt-to-descript-py"
version = "0.1.0"
description = ""
authors = ["Nelson Chen <crazysim@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
webvtt-py = "^0.4.6"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
